Use <linux/sed-opal.h> instead of outdated local copy
It appears this tool was developed during the development stages of the
in-kernel support for these interfaces but now that the IOCTLs are part
of the userspace API and <linux/sed-opal.h> exists, it seems unnecessary
to keep it around and use it.

It is also somewhat out of date in comparison to the linux one.