Makefile: Use gnu-make's LINK.o to match builtin rules
As the Makefile already uses GNUisms like $(RM) it seems fair to keep
going in that direction and use LINK.o as it's what the default %: %.o
rule uses. This includes also making the rule accept LDLIBS, although
the project doesn't make use of libraries, in case it ever does, these
would now be included by default.

I have chosen to omit including LOADLIBES (a backwards compatible alias
for LDLIBS) as it's not something anyone has heard of for many years so
it doesn't seem worth littering the Makefile with.